# docs/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Factory Ops API
  description: |
    工廠值班工作管理系統 REST API(v1.3)。
    - API-first,Web 與未來 Native App 共用同一組 API。
    - 認證採 JWT Bearer Token(access + refresh)。JWT claims 帶 `userId`、`accountName`、`rootOrgId`、`orgPath[]`、`roles[]`、`groupIds[]`、`orgManagerScopes[]`(衍生自 Organization.managerId 反查)。
    - **多租戶隔離**:所有資源(除 `Organization` 自身)歸屬於某個 `rootOrgId`,server 強制 scope。跨 root org 操作僅 `ADMIN` 系統角色可進行。
    - **時間欄位**:wire format 使用 ISO 8601 + offset(例 `2026-05-04T08:30:00+08:00`,Q-17 拍板);儲存層為 UTC `Instant`。
    - 列表端點預設使用 cursor-based pagination,支援欄位投影與增量同步。
    - **Organization 為 polymorphic 樹**(FAB → DIVISION → DEPARTMENT → SECTION),只有 leaf type 可以承載 Group / Project / Task(見 ADR-0004)。每節點有 **單一 `managerId`**(對應 ORG_MANAGER)+ **多位 `leaderIds[]`**(業務負責人列表)— 見 ADR-0010。
    - **Group 為平面結構**(屬唯一 leaf Organization),polymorphic by `type`(DEFAULT / LINE / TEAM / SHIFT…);Group 加 `settings.qa` 用於設定 QA 雙簽機制(見 ADR-0011)。
    - **Template 機制**:支援 GLOBAL(ADMIN 維護)與 ORG(各 Org 維護)兩種 scope;可 fork from GLOBAL,可直接實例化(見 ADR-0006)。
    - **跨層派工(v1.3 改寫)**:任一 Organization 節點的 manager 可直接派 ActionRequest 到其子孫中的 leaf Org(`targetOrgId` 必為 leaf);**v1.3 移除多層 relay 流程**(見 ADR-0008 v1.3)。
    - **Task QA 雙簽**:Task 建立時從所屬 Project.groupIds 對應 Group settings.qa snapshot policy;之後 Group settings 變更不影響該 Task;Review action 走 `POST /tasks/{taskId}/review`(見 ADR-0011)。
    - **HR 整合**:User 為 HR User 的本地投影,以 `accountName` 對 HR Mock REST 查詢(見 ADR-0007 v1.3 Amendment)。
...
  version: 1.3.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: 列出 Project(僅限同 orgId)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/FieldsParam'
        - $ref: '#/components/parameters/SinceParam'
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/ProjectStatus' }
        - in: query
          name: ownerId
          schema: { type: string }
        - in: query
          name: memberId
          schema: { type: string }
        - in: query
          name: groupId
          schema: { type: string }
          description: 過濾屬於某個 Group 的 Project
        - in: query
          name: tag
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: Project 清單
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectPage' }
    post:
      tags: [Projects]
      summary: 建立 Project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectRequest' }
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '422':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Projects]
      summary: 讀取單一 Project
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
          headers:
            ETag:
              schema: { type: string }
              description: 內容指紋,供條件式請求
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Projects]
      summary: 更新 Project 基本欄位
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProjectRequest' }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
    delete:
      tags: [Projects]
      summary: 軟刪除 Project
      responses:
        '204': { description: No Content }
  /projects/{projectId}/status:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    post:
      tags: [Projects]
      summary: 變更 Project 狀態
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/ProjectStatus' }
                reason: { type: string }
      responses:
        '200':
          description: 已變更
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectId}/owner:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    post:
      tags: [Projects]
      summary: 變更 Project 負責人
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newOwnerId]
              properties:
                newOwnerId: { type: string }
                reason: { type: string }
      responses:
        '200':
          description: 已變更
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
  /projects/{projectId}/members:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Memberships]
      summary: 列出 Project 成員
      responses:
        '200':
          description: 成員清單
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Membership' }
    post:
      tags: [Memberships]
      summary: 加入成員
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string }
                role: { $ref: '#/components/schemas/MembershipRole' }
      responses:
        '201':
          description: 已加入
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Membership' }
  /projects/{projectId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - $ref: '#/components/parameters/UserIdParam'
    delete:
      tags: [Memberships]
      summary: 移除成員
      responses:
        '204': { description: No Content }
  /projects/{projectId}/history:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Projects]
      summary: 列出 Project 變更歷程
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 歷程
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/HistoryEntry' }

  # ===================== Tasks =====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProjectIdParam:
      in: path
      name: projectId
      required: true
      schema: { type: string }
...
    UserIdParam:
      in: path
      name: userId
      required: true
      schema: { type: string }
    CursorParam:
      in: query
      name: cursor
      schema: { type: string }
      description: cursor-based pagination 游標
    LimitParam:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    FieldsParam:
      in: query
      name: fields
      schema: { type: string }
      description: 逗號分隔的欄位投影,例 `id,title,status,ownerId`
    SinceParam:
      in: query
      name: since
      schema: { type: string, format: date-time }
      description: 增量同步起始時間(以 updatedAt 比對)
  responses:
...
    Forbidden:
      description: 無權限
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: 找不到資源
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 業務規則衝突(狀態流、唯一性)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: 輸入驗證失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
  schemas:
...
    Problem:
      description: RFC 7807 problem details
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
      required: [title, status]
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
    PageInfo:
      type: object
      properties:
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }
    HistoryEntry:
      type: object
      properties:
        actorId: { type: string }
        action: { type: string, example: TASK_STATUS_CHANGED }
        at: { type: string, format: date-time }
        payload:
          type: object
          additionalProperties: true
...
    TimeRange:
      type: object
      properties:
        start: { type: string, format: date-time, nullable: true }
        due: { type: string, format: date-time, nullable: true }

    # ---------- Auth ----------
...
    TemplateScope:
      type: string
      enum: [GLOBAL, ORG]
      description: GLOBAL 由 ADMIN 維護,所有 Org 可見;ORG 屬某 root org
...
    MembershipRole:
      type: string
      enum: [MEMBER, LEAD, OBSERVER]
    ProjectStatus:
      type: string
      enum: [DRAFT, ACTIVE, PAUSED, COMPLETED, CANCELLED]
...
    Project:
      type: object
      properties:
        id: { type: string }
        rootOrgId: { type: string }
        organizationId: { type: string, description: 必為 leaf Organization }
        name: { type: string }
        descriptionMarkdown: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        ownerId: { type: string }
        memberIds:
          type: array
          items: { type: string }
        groupIds:
          type: array
          description: Project 所屬 Group(至少 1;MVP 必須屬同一 leaf Org)
          minItems: 1
          items: { type: string }
        schedule: { $ref: '#/components/schemas/TimeRange' }
        tags:
          type: array
          items: { type: string }
        templateRef:
          allOf:
            - $ref: '#/components/schemas/TemplateRef'
          nullable: true
          description: 若由 ProjectTemplate 實例化則記錄 snapshot
        overdue: { type: boolean, description: 衍生欄位:是否逾期(server 計算) }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
        schemaVersion: { type: integer }
      required: [id, rootOrgId, organizationId, name, status, ownerId, memberIds, groupIds]
    CreateProjectRequest:
      type: object
      required: [organizationId, name, ownerId, groupIds]
      properties:
        organizationId: { type: string, description: 必為 leaf Organization }
        name: { type: string, minLength: 1, maxLength: 120 }
        descriptionMarkdown: { type: string }
        ownerId: { type: string }
        memberIds:
          type: array
          items: { type: string }
        groupIds:
          type: array
          minItems: 1
          items: { type: string }
        schedule: { $ref: '#/components/schemas/TimeRange' }
        tags:
          type: array
          items: { type: string }
        fromTemplateId:
          type: string
          description: 若指定,從 ProjectTemplate clone(含內含 TaskTemplate 實例化)
        fromTemplateScope:
          $ref: '#/components/schemas/TemplateScope'
        fromTemplateVersion:
          type: integer
          description: 指定 Template version(預設最新 active version)
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string }
        descriptionMarkdown: { type: string }
        schedule: { $ref: '#/components/schemas/TimeRange' }
        groupIds:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
    ProjectPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Project' }
        pageInfo: { $ref: '#/components/schemas/PageInfo' }
    Membership:
      type: object
      properties:
        id: { type: string }
        rootOrgId: { type: string }
        projectId: { type: string }
        userId: { type: string }
        role: { $ref: '#/components/schemas/MembershipRole' }
        joinedAt: { type: string, format: date-time }
    TemplateRef:
      type: object
      description: 由 Template 實例化時記錄的 snapshot,稽核用,不形成執行期相依
      properties:
        templateType:
          type: string
          enum: [PROJECT_TEMPLATE, TASK_TEMPLATE]
        templateScope: { $ref: '#/components/schemas/TemplateScope' }
        templateId: { type: string }
        templateVersion: { type: integer }
        instantiatedAt: { type: string, format: date-time }

    # ---------- Task ----------
...